Replace WPF with Avalonia
