--- ArcherySimulator.Tests/ArcheryCore/ArcheryModel.cs
@@ -1,9 +1,0 @@
-namespace ArcherySimulator.Models
-{
-    public class ArcheryModel
-    {
-        public int Level { get; set; }
-        public int Experience { get; set; }
-        public int Stamina { get; set; }
-    }
-}
