+++ ArcherySimulator.Tests/ArcherySimulator.Tests.csproj
@@ -14,6 +14,11 @@
     <PackageReference Include="MSTest.TestAdapter" Version="3.0.4" />
     <PackageReference Include="MSTest.TestFramework" Version="3.0.4" />
     <PackageReference Include="coverlet.collector" Version="6.0.0" />
+    <PackageReference Include="Avalonia.Headless" Version="11.3.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\ArcherySimulator.csproj" />
   </ItemGroup>
 
 </Project>
