--- MainWindow.xaml
@@ -1,13 +1,0 @@
-﻿<Window x:Class="ArcherySimulator.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:ArcherySimulator"
-        xmlns:view="clr-namespace:ArcherySimulator.Views"
-        mc:Ignorable="d"
-        Title="ArcherySimulator" SizeToContent="WidthAndHeight" ResizeMode="NoResize">
-    <Grid>
-        <view:ArcheryView/>
-    </Grid>
-</Window>
