+++ Program.cs
@@ -1,0 +1,16 @@
+using Avalonia;
+using System;
+
+namespace ArcherySimulator;
+
+internal class Program
+{
+    public static void Main(string[] args) => BuildAvaloniaApp()
+        .StartWithClassicDesktopLifetime(args);
+
+    public static AppBuilder BuildAvaloniaApp()
+        => AppBuilder.Configure<App>()
+            .UsePlatformDetect()
+            .WithInterFont()
+            .LogToTrace();
+}
