+++ Views/ArcheryView.axaml.cs
@@ -1,0 +1,14 @@
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+using ArcherySimulator.ViewModels;
+
+namespace ArcherySimulator.Views;
+
+public partial class ArcheryView : UserControl
+{
+    public ArcheryView()
+    {
+        InitializeComponent();
+        DataContext = new ArcheryViewModel();
+    }
+}
