--- Views/ArcheryView.xaml
@@ -1,39 +1,0 @@
-﻿<UserControl x:Class="ArcherySimulator.Views.ArcheryView"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:ArcherySimulator.Views"
-             mc:Ignorable="d" 
-             d:DesignHeight="300" d:DesignWidth="300">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition/>
-            <ColumnDefinition/>
-            <ColumnDefinition/>
-            <ColumnDefinition/>
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition/>
-            <RowDefinition/>
-            <RowDefinition/>
-            <RowDefinition/>
-        </Grid.RowDefinitions>
-        
-        <Label Grid.Column="0" Grid.Row="0">Stamina</Label>
-        <TextBlock Grid.Column="1" Grid.Row="0" Text="{Binding Stamina}"/>
-        
-        <Label Grid.Column="0" Grid.Row="1">Experience</Label>
-        <TextBlock Grid.Column="1" Grid.Row="1" Text="{Binding Experience}"/>
-        
-        <Label Grid.Column="0" Grid.Row="2">Level</Label>
-        <TextBlock Grid.Column="1" Grid.Row="2" Text="{Binding Level}"/>
-        
-        <Button Grid.Column="2" Grid.Row="0" Content="Train" Command="{Binding TrainCommand}"></Button>
-        <Button Grid.Column="3" Grid.Row="0" Content="Sleep" Command="{Binding SleepCommand}"></Button>
-        <Button Grid.Column="2" Grid.Row="1" Content="Shoot" Command="{Binding ShootCommand}"></Button>
-        <Button Grid.Column="3" Grid.Row="1" Content="Break" Command="{Binding BreakCommand}" IsEnabled="{Binding BreakIsEnabled}"></Button>
-        
-        <ListBox Grid.Column="0" Grid.Row="3" Grid.ColumnSpan="4" ItemsSource="{Binding EventLog, UpdateSourceTrigger=PropertyChanged}"></ListBox>
-    </Grid>
-</UserControl>
