--- Views/ArcheryView.xaml.cs
@@ -1,30 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-using ArcherySimulator.ViewModels;
-
-namespace ArcherySimulator.Views
-{
-    /// <summary>
-    /// Interaction logic for ArcheryView.xaml
-    /// </summary>
-    public partial class ArcheryView : UserControl
-    {
-        public ArcheryView()
-        {
-            DataContext = new ArcheryViewModel();
-            InitializeComponent();
-        }
-    }
-}
